backend/requirements.txt
fastapi
uvicorn
httpx
aiosqlite
python-telegram-bot
pydantic
starlette
